% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{correlation}
\alias{correlation}
\title{Sprawdzenie korelacji x kolumn względem y kolumn}
\usage{
correlation(df, whichx, whichy, method = "spearman", high = 0.01)
}
\arguments{
\item{df}{Nasza ramka danych.}

\item{whichx}{Numery kolumn}

\item{whichy}{Numery 2 typu kolumn}

\item{method}{Metoda liczenia korelacji. Tak samo jak w funkcji bazowej cor.}

\item{high}{wartość współczynnika od jakiej mamy wypisywać pary na konsole}
}
\value{
Zwraca współczynnik korelacji dla podanych par kolumn
}
\description{
Sprawdza korelacje pomiędzy podanymi kolumnanmi
}
