% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{czas}
\alias{czas}
\title{Change date format}
\usage{
czas(
  df,
  nameFrom = "Date",
  nameTo = nameFrom,
  from = "\%Y-\%m-\%d \%H:\%M:\%S",
  to = "\%d-\%m-\%Y \%H:\%M:\%S"
)
}
\arguments{
\item{df}{Main data frame}

\item{from}{Start date format.}

\item{to}{End date format.}

\item{namesFrom}{Name of column in which we want to change date. In default "Date"}

\item{namesTo}{Name of our column at the end of process. In default we do not change name.}
}
\value{
Return a data frame with changed date format and/or column name.
}
\description{
Function changes format of date in column
}
\details{
Use to change format of time. 
  Remember to carefully copy existing format to from parameter to avoid getting all NA in column.
}
\examples{
   Examples of usage with different parameters
   czas(df, nameTo="NewDate", from="\%Y-\%d-\%m \%H:\%M:\%S",to ="\%d-\%m-\%Y \%H:\%M:\%S" )
   czas(df, to = "\%d-\%m-\%Y")
}
\references{
Gagolewski M. (2024), Deep R Programming, Melbourne
}
\seealso{
\link{wyodrebnij} for extracting specific data from date.
}
