% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{unormuj}
\alias{unormuj}
\title{Normalizacja danych numerycznych.}
\usage{
unormuj(df, which = c(2, length(colnames(df))), check = TRUE, decimals = 3)
}
\arguments{
\item{df}{Nasza ramka danych.}

\item{which}{Numery kolumn, w których mamy znormalizować dane.}

\item{check}{Czy sprawdzamy każdą kolumnę. Jeśli podajemy jakie kolumny zmienić na FALSE!. W przeciwnym wypadku funkcja sama znormalizuje tylko kolumny numeryczne.}

\item{decimals}{Ile cyfr po przecnku ma mieć wynik w kolumnach.}
}
\value{
Zwraca ramkę z znormalizowany wartościami liczbowymi.
}
\description{
Zmienia wartości wszystkich lub podanych kolumn do przedziału <0,1>.
}
