% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{czas}
\alias{czas}
\title{Zmiana typu czasu.}
\usage{
czas(
  df,
  nameFrom = "Date",
  nameTo = nameFrom,
  from = "\%Y-\%m-\%d \%H:\%M:\%S",
  to = "\%d-\%m-\%Y \%H:\%M:\%S"
)
}
\arguments{
\item{df}{Nasza ramka danych.}

\item{from}{Początkowy format daty.}

\item{to}{Końcowy format daty.}

\item{namesFrom}{Kolumna, w której posiadamy date. Domyślnie "Date".}

\item{namesTo}{Nazwa kolumny wyniku. Domyślnie nie zmieniamy nazwy.}
}
\value{
Zwraca ramkę ze zmienionym formatem danych.
}
\description{
Zmienia format czasu z x na y.
}
