% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{wyodrebnij}
\alias{wyodrebnij}
\title{Wyodrebnienie czesci daty.}
\usage{
wyodrebnij(
  df,
  name = "Date",
  what = "\%H",
  from = "\%d-\%m-\%Y \%H:\%M:\%S",
  new = TRUE,
  newName = "Hour"
)
}
\arguments{
\item{df}{Nasza ramka danych.}

\item{name}{Nazwa kolumny z datą. Domyślnie "Date".}

\item{what}{Którą część daty chcemy wyodrębnić.}

\item{from}{Format daty w  kolumnie poczatkowej.}

\item{new}{Czy tworzymy nową kolumne, czy nadpisujemy starą.}

\item{newName}{Nazwa stworzonej kolumny.}
}
\value{
Zwraca ramkę z kolumną z wyodrębnioną częścią daty.
}
\description{
Wyodrebnia podana czesc daty i tworzy badz nie nowa kolumne.
}
